Serialization refactor
Changes how save states and configurations are written to and read from disk. For instance, we use a single "Configure" function for each gameboy class component that takes a functor as an argument, which then either writes or reads to the config file, depending on what type the functor is.